requirements-x86.txt
# BCM v7 — x86 development/simulation dependencies
flask>=3.0
flask-sock>=0.7
Pillow>=10.0
